update main timeline late phase
Mentions of events after returner's fate, which are the result of the Storyline-Supermerge mini-Project. (which currently attempts to merge the CW Storyline with the Panzerdampf Storyline.
While the migration event would fit nicely into the CW-Storyline imo, the destruction of the islands (you could also say "narrative cleanup") isn't necessary for the CW Storyline and might leave the reader with a stale feeling, so mayybe it would be better to only mention it in the origins of Panzerdampf.